+++ src/components/nabla/Pools/Backstop/AddLiquidity/useAddLiquidity.ts
@@ -22,7 +22,9 @@ export const useAddLiquidity = (poolAddress: string, tokenAddress: string) => {
 
   const form = useForm<AddLiquidityValues>({
     resolver: yupResolver(schema),
-    defaultValues: {},
+    defaultValues: {
+      amount: undefined,
+    },
   });
 
   const mutation = useContractWrite({
@@ -40,7 +42,7 @@ export const useAddLiquidity = (poolAddress: string, tokenAddress: string) => {
     },
   });
 
-  const onSubmit = form.handleSubmit((variables: AddLiquidityValues) =>
+  const onSubmit = form.handleSubmit((variables) =>
     mutation.mutate([decimalToNative(variables.amount, defaultDecimals).toString()]),
   );
