+++ src/components/nabla/Swap/To/index.tsx
@@ -7,10 +7,10 @@ import pendulumIcon from '../../../../assets/pendulum-icon.svg';
 import { config } from '../../../../config';
 import { defaultDecimals } from '../../../../config/apps/nabla';
 import { subtractPercentage } from '../../../../helpers/calc';
+import { useTokenOutAmount } from '../../../../hooks/nabla/useTokenOutAmount';
 import { TokensData } from '../../../../hooks/nabla/useTokens';
 import useBoolean from '../../../../hooks/useBoolean';
 import { useDebouncedValue } from '../../../../hooks/useDebouncedValue';
-import { useTokenOutAmount } from '../../../../hooks/useTokenOutAmount';
 import { nativeToDecimal, prettyNumbers, roundNumber } from '../../../../shared/parseNumbers';
 import Balance from '../../../Balance';
 import { numberLoader } from '../../../Loader';
